rename folder to legojs
