+++ components/Alert/Alert.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Brick from 'lego-components'
+import Brick from 'legojs'
 import Styles from './Alert.module.css' 
 
 const Alert = ({as='div', ...props}) => {
